fix defects which issues are produced when app runs on Nexus9

+++ res/raw/pfl_color_30_fs.glsl
@@ -12,7 +12,7 @@ layout( location = 0) out vec4 fragColor;
 
 uniform highp mat3 uNormalMatrix;
 
-uniform int uLightState[8];
+uniform lowp int uLightState[8];
 
 struct LightInfo {
     highp vec4 ambient;
